Moved array pools and changed namespace.
