+++ GDX/Runtime/Collections/Pooling/ManagedArrayPool.cs.meta
