+++ GDX/Runtime/Collections/Pooling/ManagedArrayPool_PowerOfTwoLengths.cs
@@ -1,7 +1,7 @@
 // dotBunny licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace GDX.Collections
+namespace GDX.Collections.Pooling
 {
     public struct ManagedArrayPool_PowerOfTwoLengths<T>
     {
