+++ GDX/Runtime/Collections/Pooling/ManagedArrayPool_PowerOfTwoLengths.cs.meta
